Extract shared UnityHandlerScope helper for tests
Both UnityExceptionIntegrationTests and WebGLExceptionIntegrationTests
defined an identical HandlerScope class to swap and restore
Debug.unityLogger.logHandler. Extract it to a shared UnityHandlerScope
file so future changes to the seam are a single edit.
